analyze.py does analysis on the generated python files. You just need to specify how many runs to do and how big of graphs you want to generate. and what fitness function to use
analysis.py is where all of the functions are that acutally analyze the graphs

I will be adding more analysis stuff soon that does trends so we can show how the analysis stuff changes as the size of the graph changes

I added the fitnessFunction module. This is where we will store the fitness functions that will be used in evolution and analysis. There is an added field to the settings file to
allow for specifying the fitness function.

I also added relative selection operators, not sure if that was in the last push or not. But I did that too.

And now for a bad joke:

Why can't a nose be 12 inches long?

--Because then it'd be a foot!